refactor: transactions observer ckBTC and Icrc prefixes
